refactor: add Header-component
