+++ src/components/containers/domains/headers/Header.js
@@ -1,0 +1,11 @@
+import { View } from "react-native-web";
+import Title from "./Title";
+import SubTitle from "./SubTitle";
+
+
+export default function Header() {
+    return <View>
+        <Title />
+        <SubTitle />
+    </View>
+}
