+++ src/components/containers/domains/headers/Title.js
@@ -1,5 +1,5 @@
 import { View } from 'react-native';
-import theme from '../theme';
+import theme from '../../../../theme';
 
 export default function Title() {
